Attempted to fix `fmt` problem
No way to properly test except with reverting the changes to use Canvas API, unless a test is written to check the string representation.

+++ src/lib.rs
@@ -157,7 +157,7 @@ impl fmt::Display for Universe {
         for line in self.cells.as_slice().chunks(self.width as usize) {
             for &cell in line {
                 // let symbol = if cell == Cell::Dead { '◻' } else { '◼' };
-                let symbol = if cell == false { '◻' } else { '◼' };
+                let symbol = if cell == false as usize { '◻' } else { '◼' };
                 write!(f, "{}", symbol)?;
             }
             write!(f, "\n")?;
